+++ lib/internet_state_manager.dart
@@ -4,5 +4,6 @@ export 'src/widgets/internet_state_manager_widget.dart';
 export 'src/utils/internet_state_options.dart';
 export 'src/internet_state_manager_app.dart';
 export 'src/utils/enums/internet_state_enum.dart';
+export 'src/utils/extensions/context_extension.dart';